@@ -2,7 +2,5 @@ import { defineConfig } from 'vite';
 import solidPlugin from 'vite-plugin-solid';
 
 export default defineConfig({
-  plugins: [
-    solidPlugin(),
-  ],
+  plugins: [solidPlugin()],
 });
